Refactored Common Widgets
