+++ lib/src/features/authentication/presentation/pages/auth_page.dart
@@ -1,4 +1,4 @@
-import 'package:apk_iman_ba/components/custom_bnb.dart';
+import 'package:apk_iman_ba/src/shared/common_widgets/custom_bnb.dart';
 import 'package:apk_iman_ba/src/features/authentication/presentation/pages/login_page.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
