+++ lib/src/features/questions/presentation/pages/waiting_list_page.dart
@@ -1,5 +1,5 @@
 import 'package:apk_iman_ba/Services/database_service.dart';
-import 'package:apk_iman_ba/components/custom_listview.dart';
+import 'package:apk_iman_ba/src/shared/common_widgets/custom_listview.dart';
 import 'package:apk_iman_ba/src/features/questions/domain/models/question_model.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/foundation.dart';
