+++ lib/src/features/search/presentation/pages/search_page.dart
@@ -1,4 +1,4 @@
-import 'package:apk_iman_ba/components/custom_shimmer.dart';
+import 'package:apk_iman_ba/src/shared/common_widgets/custom_shimmer.dart';
 import 'package:apk_iman_ba/src/features/questions/domain/models/question_model.dart';
 import 'package:apk_iman_ba/services/database_service.dart';
 import 'package:firebase_database/firebase_database.dart';
