+++ lib/src/shared/common_widgets/custom_listview.dart
@@ -1,4 +1,4 @@
-import "package:apk_iman_ba/components/custom_shimmer.dart";
+import "package:apk_iman_ba/src/shared/common_widgets/custom_shimmer.dart";
 import "package:apk_iman_ba/src/features/questions/domain/models/question_model.dart";
 import "package:apk_iman_ba/src/features/questions/presentation/pages/details_page.dart";
 import 'package:apk_iman_ba/src/features/questions/presentation/pages/wait_answer_page.dart';
